feat: add AppLifecycleUI
